pyproject.toml
[project]
name = "lib_layered_config"
version = "0.0.1"
description = "Cross-platform layered configuration loader for Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "rich-click>=1.9.1",
  "lib_cli_exit_tools>=1.3.1",
]
license = { text = "MIT" }
authors = [{ name = "bitranox", email = "bitranox@gmail.com" }]
keywords = ["configuration", "layered", "dotenv", "env", "toml"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/bitranox/lib_layered_config"
Repository = "https://github.com/bitranox/lib_layered_config.git"
Issues = "https://github.com/bitranox/lib_layered_config/issues"

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "pyright>=1.1",
  "bandit>=1.7",
  "pip-audit>=2.7",
  "hypothesis>=6.112",
  "coverage[toml]>=7.6",
  "mypy>=1.10",
]
yaml = [
  "PyYAML>=6.0",
]

[project.scripts]
lib_layered_config = "lib_layered_config.cli:main"
lib-layered-config = "lib_layered_config.cli:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"



[tool.hatch.build.targets.wheel]
packages = ["src/lib_layered_config"]
include = ["src/lib_layered_config/py.typed"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint.per-file-ignores]
# ignore the "unused import sys" in the jupyter notebook - which we need (it is a linter hickup)
"notebooks/*.ipynb" = ["F401"]

[tool.pyright]
typeCheckingMode = "strict"
exclude = [
  "scripts/menu.py",
]

[tool.pytest.ini_options]
addopts = ["--maxfail=1", "--cov=lib_layered_config", "--cov-report=term-missing", "--cov-fail-under=90", "--doctest-modules"]
doctest_optionflags = ["ELLIPSIS", "NORMALIZE_WHITESPACE"]

[tool.coverage.run]
branch = true
parallel = false
data_file = ".coverage.unit"
source = ["src/lib_layered_config"]

[tool.coverage.report]
omit = [
  "tests/*",
  "packaging/*",
]
fail_under = 75
show_missing = true

[tool.importlinter]
root_package = "lib_layered_config"

[[tool.importlinter.contracts]]
name = "Clean architecture layers"
type = "layers"
layers = [
  "lib_layered_config.domain",
  "lib_layered_config.application",
  "lib_layered_config.adapters",
  "lib_layered_config",
]

[[tool.importlinter.contracts]]
name = "No adapters in domain"
type = "forbidden"
modules = ["lib_layered_config.domain"]
forbidden_modules = ["lib_layered_config.adapters", "lib_layered_config.application", "lib_layered_config.core"]

